Network.TLS.SessionManager

Description

In-memory TLS session manager.

Limitation: you can set the maximum size of the session data database.

Automatic pruning: old session data over their lifetime are pruned automatically.

Energy saving: no dedicate pruning thread is running when the size of session data database is zero.

(Replay resistance: each session data is used at most once to prevent replay attacks against 0RTT early data of TLS 1.3.)

Documentation

data Config

Configuration for session managers.

Constructors

Config

Fields

ticketLifetime :: !Int

Ticket lifetime in seconds.

pruningDelay :: !Int

Pruning delay in seconds. This is set to reaperDelay.

dbMaxSize :: !Int

The limit size of session data entries.

defaultConfig :: Config

Lifetime: 1 day , delay: 10 minutes, max size: 1000 entries.

newSessionManager :: Config -> IO SessionManager

Creating an in-memory session manager.